requirements.txt
numpy
pandas
scikit-learn
seaborn
matplotlib
openpyxl
fpdf2
